backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# 3D Processing
open3d==0.18.0
trimesh==4.0.8
numpy==1.26.3
scipy==1.11.4

# Feature extraction and similarity search
scikit-learn==1.4.0
faiss-cpu==1.7.4

# File handling
python-multipart==0.0.6
aiofiles==23.2.1
python-magic==0.4.27

# Utilities
pydantic-extra-types==2.3.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Monitoring
python-json-logger==2.0.7
